Refactor venue routing and preview
