+++ .vscode/settings.json
@@ -17,5 +17,6 @@
   "[json]": {
     "editor.tabSize": 2
   },
-  "typescript.tsdk": "./node_modules/typescript/lib"
+  "typescript.tsdk": "./node_modules/typescript/lib",
+  "editor.tabSize": 2
 }
