+++ functions/auth.js
@@ -1,7 +1,7 @@
 const functions = require("firebase-functions");
 const PROJECT_ID = functions.config().project.id;
 
-exports.checkAuth = (context) => {
+exports.checkAuth = (venueId, context) => {
   if (!context.auth || !context.auth.token) {
     throw new functions.https.HttpsError("unauthenticated", "Please log in");
   }
