+++ functions/venue.js
@@ -1,6 +1,7 @@
 const admin = require("firebase-admin");
 const functions = require("firebase-functions");
 const { checkAuth } = require("./auth");
+const { HttpsError } = require("firebase-functions/lib/providers/https");
 const PROJECT_ID = functions.config().project.id;
 
 // interface Question {
@@ -51,12 +52,12 @@ const createVenueData = (data, context) => ({
   code_of_conduct_questions: [
     {
       name: "contributeToExperience",
-      text: "Are you willing ton contribute to the experience?",
+      text: "Are you willing to contribute to the experience?",
     },
   ],
   config: {
     landingPageConfig: {
-      checkList: ["Enjoy our amazing venue"],
+      checkList: [],
       coverImageUrl: data.bannerImageUrl,
       joinButtonText: "Enter our venue",
       subtitle: data.tagline,
@@ -64,7 +65,7 @@ const createVenueData = (data, context) => ({
     },
   },
   presentation: [data.longDescription],
-  quotations: [{ author: "Max", text: "'Bullish on life' Bull" }],
+  quotations: [],
   videoIframeUrl: "https://www.youtube.com/watch?v=dQw4w9WgXcQ",
   theme: {
     primaryColor: "#bc271a",
@@ -82,33 +83,58 @@ const createVenueData = (data, context) => ({
   //@debt need to do something with mapIconUrl
 });
 
+const getVenueId = (name) => {
+  return name.replace(/\W/g, "").toLowerCase();
+};
+
+const checkUserIsOwner = async (venueId, uid) => {
+  await admin
+    .firestore()
+    .collection("venues")
+    .doc(venueId)
+    .get()
+    .then((doc) => {
+      if (!doc.exists()) {
+        throw new HttpsError("not-found", `Venue ${venueId} does not exist`);
+      }
+      const venue = doc.data();
+      if (!doc.owners || !doc.owners.includes(uid)) {
+        throw new HttpsError(
+          "permission-denied",
+          `User is not an owner of ${venueId}`
+        );
+      }
+    })
+    .catch((err) => {
+      throw new HttpsError(
+        "internal",
+        `Error occurred obtaining venue ${venueId}: ${err.toString()}`
+      );
+    });
+};
+
 exports.createVenue = functions.https.onCall(async (data, context) => {
   checkAuth(context);
 
   // @debt this should be typed
   const venueData = createVenueData(data, context);
+  const venueId = getVenueId(data.name);
 
-  await admin
-    .firestore()
-    .collection("venues")
-    .doc(data.name.replace(/\W/g, "").toLowerCase())
-    .set(venueData);
+  await admin.firestore().collection("venues").doc(venueId).set(venueData);
 
   return venueData;
 });
 
 exports.updateVenue = functions.https.onCall(async (data, context) => {
-  console.log("HELLO");
-  checkAuth(context);
+  checkAuth(venueId, context);
+  const venueId = getVenueId(data.name);
+
+  checkUserIsOwner(venueId, context.auth.token.user_id);
 
   // @debt this should be typed
   const venueData = createVenueData(data, context);
 
-  await admin
-    .firestore()
-    .collection("venues")
-    .doc(data.name.replace(/\W/g, ""))
-    .update(venueData);
+  await admin.firestore().collection("venues").doc(venueId).update(venueData);
 
   return venueData;
 });
