+++ src/components/organisms/TemplateRouter/index.ts
@@ -1,1 +1,1 @@
-export { default } from "./TemplateRouter";
+// export { default } from "./TemplateRouter";
