--- src/components/templates/EntranceExperience/index.ts
@@ -1,1 +1,0 @@
-export * from "./EntranceExperience";
