+++ src/pages/VenueLandingPage/VenueLandingPage.scss
@@ -25,11 +25,11 @@
         width: 150px;
         height: 150px;
         object-fit: cover;
-        border-radius: 75px;
+        border-radius: 28px;
         overflow: hidden;
       }
 
-      .playa-marketing-preview {
+      .playa-venue-preview {
         position: absolute;
         top: -200px;
         right: -600px;
