+++ src/types/VenueTemplate.ts
@@ -5,7 +5,7 @@ export enum VenueTemplate {
   jazzbar = "jazzbar",
   friendship = "friendship",
   partymap = "partymap",
-  artPiece = "artpiece",
+  artpiece = "artpiece",
   camp = "camp",
 }
 
@@ -16,6 +16,6 @@ export const venueTemplateToTemplateType: Record<
   [VenueTemplate.jazzbar]: "ZOOM_ROOM",
   [VenueTemplate.friendship]: "ZOOM_ROOM",
   [VenueTemplate.partymap]: "THEME_CAMP",
-  [VenueTemplate.artPiece]: "ZOOM_ROOM",
+  [VenueTemplate.artpiece]: "ZOOM_ROOM",
   [VenueTemplate.camp]: "THEME_CAMP",
 };
